Implement on-the-fly addition/removal of entities
