+++ homeassistant/components/nasweb/manifest.json
@@ -8,7 +8,7 @@
   "homekit": {},
   "integration_type": "hub",
   "iot_class": "local_push",
-  "requirements": ["webio-api==0.1.3"],
+  "requirements": ["webio-api==0.1.4"],
   "ssdp": [],
   "zeroconf": []
 }
