+++ requirements_all.txt
@@ -2915,7 +2915,7 @@ weatherflow4py==0.2.21
 webmin-xmlrpc==0.0.2
 
 # homeassistant.components.nasweb
-webio-api==0.1.3
+webio-api==0.1.4
 
 # homeassistant.components.whirlpool
 whirlpool-sixth-sense==0.18.8
