+++ requirements_test_all.txt
@@ -2298,7 +2298,7 @@ watchdog==2.3.1
 weatherflow4py==0.2.21
 
 # homeassistant.components.nasweb
-webio-api==0.1.3
+webio-api==0.1.4
 
 # homeassistant.components.webmin
 webmin-xmlrpc==0.0.2
